+++ SandboxiePlus/SandMan/SandMan.cpp
@@ -2424,13 +2424,41 @@ void CSandMan::OnStatusChanged()
 		}
 
 		int DynData = theAPI->IsDyndataActive();
-		if (DynData == 0)
-			QMessageBox::critical(this, "Sandboxie-Plus", tr("Your Windows build exceeds the current support capabilities of your Sandboxie version, "
-				"resulting in the disabling of token-based security isolation. Consequently, all applications will operate in application compartment mode without secure isolation.\r\n"
-				"Please check if there is an update for sandboxie."));
-		else if (DynData == -1)
-			OnLogMessage(tr("Your Windows build exceeds the current known support capabilities of your Sandboxie version, "
-				"Sandboxie will attempt to use the last-known offsets which may cause system instability."));
+		if (DynData != 1) 
+		{
+			RTL_OSVERSIONINFOEXW versionInfo;
+			memset(&versionInfo, 0, sizeof(RTL_OSVERSIONINFOEXW));
+			versionInfo.dwOSVersionInfoSize = sizeof(RTL_OSVERSIONINFOEXW);
+			NTSTATUS(WINAPI *RtlGetVersion)(PRTL_OSVERSIONINFOEXW);
+			*(void**)&RtlGetVersion = GetProcAddress(GetModuleHandleA("ntdll.dll"), "RtlGetVersion");
+			if (RtlGetVersion != NULL) 
+				RtlGetVersion(&versionInfo);
+			else
+				GetVersionExW((LPOSVERSIONINFOW)&versionInfo); // since windows 10 this one is lying
+			RtlGetVersion(&versionInfo);
+
+			if (DynData == 0) 
+			{
+				QString Message = tr("Your Windows build %1 exceeds the current support capabilities of your Sandboxie version, "
+					"resulting in the disabling of token-based security isolation. Consequently, all applications will operate in application compartment mode without secure isolation.\r\n"
+					"Please check if there is an update for sandboxie.").arg(versionInfo.dwBuildNumber);
+				OnLogMessage(Message, true);
+
+				int IgnoreUnkBuild = theConf->GetInt("Options/IgnoreUnkBuild", 0);
+				if (IgnoreUnkBuild != versionInfo.dwBuildNumber)
+				{
+					bool Ignore = false;
+					CCheckableMessageBox::question(this, "Sandboxie-Plus", Message, tr("Don't show this message again for the current build."), &Ignore, QDialogButtonBox::Ok, QDialogButtonBox::Ok, QMessageBox::Critical);
+					if (Ignore)
+						theConf->SetValue("Options/IgnoreUnkBuild", (int)versionInfo.dwBuildNumber);
+				}
+			}
+			else if (DynData == -1)
+			{
+				OnLogMessage(tr("Your Windows build %1 exceeds the current known support capabilities of your Sandboxie version, "
+					"Sandboxie will attempt to use the last-known offsets which may cause system instability.").arg(versionInfo.dwBuildNumber), true);
+			}
+		}
 
 		if (isVisible())
 			CheckSupport();
@@ -3562,6 +3590,8 @@ void CSandMan::OnResetMsgs()
 		theConf->DelValue("Options/WarnOpenCOM");
 
 		theConf->DelValue("Options/WarnWizardOnClose");
+
+		theConf->DelValue("Options/IgnoreUnkBuild");
 	}
 
 	theAPI->GetUserSettings()->UpdateTextList("SbieCtrl_HideMessage", QStringList(), true);